common: Move the EncryptedData structure from common to the library
Move the EncryptedData from the common.rs file to the Keylime library in
the keylime::crypto::encrypted_data module.
