+++ keylime-agent/src/keys_handler.rs
@@ -2,7 +2,7 @@
 // Copyright 2021 Keylime Authors
 
 use crate::{
-    common::{EncryptedData, JsonWrapper},
+    common::JsonWrapper,
     config::KeylimeConfig,
     payloads::{Payload, PayloadMessage},
     Error, QuoteData, Result,
@@ -11,8 +11,9 @@ use actix_web::{http, web, HttpRequest, HttpResponse, Responder};
 use base64::{engine::general_purpose, Engine as _};
 use keylime::crypto::{
     self,
-    symmkey::{KeySet, SymmKey},
     auth_tag::AuthTag,
+    encrypted_data::EncryptedData,
+    symmkey::{KeySet, SymmKey},
 };
 use log::*;
 use serde::{Deserialize, Serialize};
